sqlc migration (#695)
* Moved all sql calls to use sqlc
* Removed sqlx. removed all reflection outside of tests
* added benchmark logic to db UpdateFullGame
* Changed language around create/update to just be "save"
* added explain tests to ensure queries are using indexes where possible
* removed rules from games table.
* Migrated tables to snake case for more consistent naming. Also using NOT NULL and default values to better conform to serializing go models.
* calling vacuum if the db is upgraded.
* fixed guest joining game workflow
* Changed all instances of MineField to Minefield (and all text/db entries on upgrade)
* Changed all mine-field strings to minefield
* fixed in game tech page not showing up
* added games list --full --generate for ensuring all games load from the DB, can generate a turn
* fixed issue with by hand transfers not accounting for existing cargo when transferring
* fixed cargo transfer dest capacity to work with infinite capacity
* improved by hand transfer logs
* fixed split fleet cargo transfer issue where a split fleet created a transfer to/from itself.
added test for scenario of splitting off a colony ship from the main fleet but keeping colonists
* fixed issues with splitting into new fleets but deleting old fleets
* preventing empty cargo transfers from being created
* fixed add duplicate waypoint bug

+++ frontend/src/routes/(user)/games/archived/+page.svelte
@@ -68,14 +68,17 @@
 	let descending = $state(true);
 
 	async function archiveGame(game: Game) {
-		if (confirm(`Are you sure you want to unarchive ${game.name}?`)) {
+		if (game.id && confirm(`Are you sure you want to unarchive ${game.name}?`)) {
 			await PlayerService.unArchiveGame(game.id);
 			games = games.filter((g) => g.id !== game.id);
 		}
 	}
 
 	async function deleteGame(game: Game) {
-		if (confirm(`Are you sure you want to delete ${game.name}? This operation cannot be undone.`)) {
+		if (
+			game.id &&
+			confirm(`Are you sure you want to delete ${game.name}? This operation cannot be undone.`)
+		) {
 			await GameService.deleteGame(game.id);
 			games = games.filter((g) => g.id !== game.id);
 		}
@@ -124,9 +127,9 @@
 				{#if column.key == 'name'}
 					<a class="cs-link text-xl" href="/games/{row.id}">{cell}</a>
 				{:else if column.key == 'createdAt'}
-					{format(parseJSON(row.createdAt), 'E, MMM do yyyy hh:mm aaa')}
+					{format(parseJSON(row.createdAt ?? ''), 'E, MMM do yyyy hh:mm aaa')}
 				{:else if column.key == 'updatedAt'}
-					{format(parseJSON(row.updatedAt), 'E, MMM do yyyy hh:mm aaa')}
+					{format(parseJSON(row.updatedAt ?? ''), 'E, MMM do yyyy hh:mm aaa')}
 				{:else if column.key == 'hostId'}
 					{row.players.find((p) => p.userId === row.hostId)?.name}
 				{:else if column.key == 'players'}
